:truck: Move : guest 패키지 생성 및 관련 파일 이동
<footer>
- 관련: #272
